factor repetitive code
