+++ src/components/encoding-pane/range-component.scss
@@ -1,0 +1,41 @@
+$border: 1px solid #999;
+$radius: 3px;
+
+.range-filter-pane {
+  width: 80%;
+  margin-left: auto;
+  margin-right: auto;
+  margin: 10px auto;
+
+  input {
+    font-size: 11px;
+    border: 0;
+    outline: 0;
+    background: transparent;
+  }
+
+  input:focus, input:hover {
+    border-bottom: $border
+  }
+
+  a {
+    display: inline;
+  }
+
+  :global .fa {
+    &.fa-pencil {
+      cursor: default;
+      padding: 2px 3px 1px;
+      border-radius: $radius;
+
+      &:hover {
+        color: #fff;
+        background-color: #999;
+      }
+    }
+  }
+}
+
+.bound {
+  font-size: 10px;
+}
