+++ src/components/encoding-pane/range-filter-shelf.scss
@@ -1,52 +1,13 @@
-$border: 1px solid #999;
-$radius: 3px;
+@import './range-component.scss';
+
 $track-color: #1f77b4;
 $active-color: #1f9cb4;
-
-.range-filter-pane {
-  width: 80%;
-  margin-left: auto;
-  margin-right: auto;
-  margin: 10px auto;
-
-  input {
-    font-size: 11px;
-    border: 0;
-    outline: 0;
-    background: transparent;
-  }
-
-  input:focus, input:hover {
-    border-bottom: $border
-  }
-
-  a {
-    display: inline;
-  }
-
-  :global .fa {
-    &.fa-pencil {
-      cursor: default;
-      padding: 2px 3px 1px;
-      border-radius: $radius;
-
-      &:hover {
-        color: #fff;
-        background-color: #999;
-      }
-    }
-  }
-}
 
 .date-time-picker-wrapper {
   padding: 5px;
   background-color: #eee;
   border: 1px solid #eee;
   border-radius: 3px;
-}
-
-.bound {
-  font-size: 10px;
 }
 
 :global {
