+++ src/components/encoding-pane/time-unit-filter-shelf.scss
@@ -1,4 +1,5 @@
-@import './range-filter-shelf.scss';
+@import './range-component.scss';
+
 .time-unit-changer {
   padding: 5px;
   background-color: #eee;
